Complete project and merge test to main
Moved all the test files out to the main directory. Made the output look
almost identical to the provided image. Not sure how the page is
supposed to behave in regards to sizing because the solution is an image
not a webpage. All reference borders have been commented out as they are
no longer needed.